+++ Tests/Get-RubrikEvent.Tests.ps1
@@ -26,26 +26,31 @@ Describe -Name 'Public/Get-RubrikEvent' -Tag 'Public', 'Get-RubrikEvent' -Fixtur
                 'name'                   = 'VirtualMachine01'
                 'id'                     = 'VirtualMachine:::11111'
                 'eventType'              = 'Replication'
+                'time'                   = 'Mon Aug 10 07:12:14 UTC 2019'
             },
             @{ 
                 'name'                   = 'VirtualMachine02'
                 'id'                     = 'VirtualMachine:::22222'
                 'eventType'              = 'Backup'
+                'time'                   = 'Mon Aug 11 07:12:14 UTC 2019'
             },
             @{ 
                 'name'                   = 'VirtualMachine03'
                 'id'                     = 'VirtualMachine:::33333'
                 'eventType'              = 'CloudNativeSource'
+                'time'                   = 'Mon Aug 12 07:12:14 UTC 2019'
             },
             @{ 
                 'name'                   = 'VirtualMachine04'
                 'id'                     = 'VirtualMachine:::44444'
                 'eventType'              = 'Replication'
+                'time'                   = 'Mon Aug 13 07:12:14 UTC 2019'
             },
             @{ 
                 'name'                   = 'VirtualMachine05'
                 'id'                     = 'VirtualMachine:::55555'
                 'eventType'              = 'Replication'
+                'time'                   = 'Mon Aug 14 07:12:14 UTC 2019'
             }
         }
         It -Name 'Should Return count of 5' -Test {
@@ -56,9 +61,13 @@ Describe -Name 'Public/Get-RubrikEvent' -Tag 'Public', 'Get-RubrikEvent' -Fixtur
             {Get-RubrikEvent -Name doesnotexist -ErrorAction Stop} |
                 Should -Throw
         }
-
+        It -Name 'Date property should be a datetime object' -Test {
+            (Get-RubrikEvent)[1].Date |
+                Should -BeOfType DateTime
+        }
+        
         Assert-VerifiableMock
-        Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Times 1
-        Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik' -Times 1
+        Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Exactly 2
+        Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik' -Exactly 2
     }
 }